Update README to show correct form for tag in 'ref'
This is a subtle side-effect of 6ad8f6d0d8.

Previously, `ref` from `artifacts.yml` was being passed to
`project.commits.get()`.  For a tag, you could pass *either* `v1.2.3` or
`tags/v1.2.3`. This is probably because Git itself allows either form
(e.g. `git show`). The latter form was specified by `README.md`.

Now, `ref` is being passed to `project.pipelines.list()`, which *does
not* allow the `tags/v1.2.3` form. This is probably because this hits
the GitLab database, and not the git repo.

So the end result is that with this change, the `tags/v1.2.3` form is no longer acceptable.