Improve error when update fails due to job token
The job token doesn't provide access to the API endpoints required to
perform the update command. If you try it anyway, GitLab returns a 404
status code when accessing a project via the API using a job token. Work
around this by checking the type of token during update and print a
reasonable error message.